refactor: Move public FontFamilyHelper
